html/semantics/scripting-1/the-script-element/css-module/valid-content-type.tentative.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>JSON modules: Content-Type</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
function check(t, v) {
  t.step(() => {
    assert_equals(typeof v, "object");
    t.done();
  });
}
const t1 = async_test("text/css");
</script>
<script type="module" onerror="t1.step(() => assert_unreached(event))">
import v from "../serve-with-content-type.py?fn=css-module/resources/utf8.css&ct=text/css";
check(t1, v);
</script>
